week-0/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
openai>=1.0
pydantic>=2.0
tiktoken>=0.6
google-genai>=1.0
